corrected three typos
I've corrected three typos; the article is about the alternative box-model, not the alternate box model. There is one more typo at line 170, though (still alternate instead of alternative); however, being that the name of a class, and since an exercise is embed right after, I'm afraid that changing it may cause problems.