fix: tiny text in syntaxboxes
This removes the use of the `syntaxbox` class and replaces it with correctly using the relevant brush.
I have also taken the opportunity to ensure all `pre` elements have the `notranslate` class.
I have also cleaned up instances where `pre` elements was wrapped in `div` elements unnecessarily as well as other instances where `syntaxbox` was used for content that is clearly not syntax.

fix #1123