Drop <main role=main> example; refine explanation
We shouldn’t use the word “recommended” to describe using
<main role=main> — because the ARIA in HTML spec actually normatively
recommends against it. So instead, we should just explain that IE 11 and
earlier don’t recognize <main> but do recognize role=main, and leave it
up to the reader to decide if they want to use it in spite of the spec
recommending against it.

Fixes https://github.com/mdn/sprints/issues/3352