Drop ariaRelevant spec table (not in ARIA 1.2)
Reflection of the aria-relevant content attribute to the ariaRelevant
IDL attribute has been dropped from ARIA 1.2:

https://github.com/w3c/aria/issues/1267

> The prose of aria-relevant includes some special rules that cannot be
> expressed in standard WebIDL, so I'm removing its DOMString-based
> reflection in ARIA 1.2. … I think we should be able to bring this back
> in 1.3.

https://www.w3.org/TR/wai-aria-1.2/ (last published 18 December 2019) is
not up to date with that change, but https://w3c.github.io/aria/ (last
published 22 February 2021) is — and it doesn’t have ariaRelevant.

So, we should drop the Specifications table from the ariaRelevant MDN
article for now.