Clarify: lastIndex is RegExp instance property
This change updates the RegExp article to correct a misclassification of
lastIndex as a static property; it’s not a static property, but instead
an instance property.

However, it’s not a property of RegExp.prototype but instead is only exposed
from RegExp instances — so this change adds some special-casing for that in
the article for lastIndex itself, as well as in the RegExp article.

Fixes https://github.com/mdn/content/issues/2708