use const instead of let declaration for some vars (#2912)
In the live version of the sequencer this page describes how to create, there exists a commit that does the same thing as this commit does for this page, except it does it for all applicable variable declarations, instead of the subset I verified on this page. There are probably more variable declarations that can use `const` instead of `let` on this page.

See https://github.com/mdn/webaudio-examples/commit/1217eb894d4a1865058939df211e4209bcb9205d.