csp: script-src: examples: avoid id collision
Avoids an id/anchor collision between Example's and Sources's strict-dynamic sections.

This document embeds the Sources section from `Web/HTTP/Headers/Content-Security-Policy/default-src` which already sets a `strict-dynamic` id, and both end up in the resulting page[1]. Consequently, the `strict-dynamic` anchor in this document here - the second one in the DOM - was not linkable as desired.

This change is a bit desperate as it is only a manual solution for the one problem I spotted. A systemic solution would be preferred. I can envision e.g. consistently tagging the sources - so all of them become linkable - and doing so with a "sources" prefix to limit the likelihood of collisions but that would break existing links/bookmarks. I assume this is not the first time a problem like this popped up in a large project like this and maybe someone has a goto solution which can be applied as a follow-up.

[1]: https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Content-Security-Policy/script-src