It is very confusing
> The json() method of the Body mixin takes a Response stream and reads it to completion. It returns a promise that resolves with the result of parsing the body text as JSON.

can read like "It returns a promise that resolves ... as JSON"

but in fact, it means "It returns a promise that resolves with (the result of parsing the body text as JSON)"

So it is: by parsing the body text as JSON, return the result in the form of a JavaScript object.

It is mentioned as such down below the docs, but the title can give people wrong impressions that it is returning "JSON".

There are so many places where people think it is "JSON", when it is JavaScript object.

https://github.com/mdn/fetch-examples/blob/e59e035214d21926eb7fc474e0cea88b33199b00/fetch-json/index.html

https://jsonplaceholder.typicode.com/

```
fetch('https://jsonplaceholder.typicode.com/todos/1')
  .then(response => response.json())
  .then(json => console.log(json))
```

You see, we have a string `s`, we use `parseInt(s)` to get back and integer.  We would not call that integer as a string, would we?

So, we have JSON, we use `.json()` to parse it and get back a JavaScript object, and we call this JavaScript object a "JSON".  That is very wrong.