Use “BigInt values” to clarify they’re not objects
This change updates the BigInt article to clarify that code does not
operate on BigInt objects but instead on bigint primitive values. The
ECMAScript spec uses the term “BigInt value” to refer to bigint
primitives — and the shorthand “BigInt” to refer to BigInt values — and
normatively defines what a BigInt value is:

https://tc39.es/ecma262/#sec-terms-and-definitions-bigint-value

> 4.4.28 BigInt value
> primitive value corresponding to an arbitrary-precision integer value

…so this change follows that, and updates the BigInt article to
consistently use “BigInt value”, as normal type — non-monospaced,
without `<code></code>` markup — to make clear that what’s being
operated on in code is not a BigInt *object*, but a primitive value.

Fixes https://github.com/mdn/content/issues/3336