Add SeeCompatTable to “Preloading content” doc (#3428)
* Push the “fixable flaws” button

* Add SeeCompatTable to “Preloading content” doc

This change adds SeeCompatTable to the “Preloading content with
rel="preload"” article — because there are some values for the “as”
attribute which aren’t yet supported in Chrome and other browsers. And
the unsupported values include `audio`, `video`, and `document` — which
are resources that developers somewhat naturally expect to be supported
but aren’t. So we should direct them to the Compat table for details.

Related: https://github.com/mdn/browser-compat-data/issues/9577