Drop unnecessary terminology from postMessage docs
https://html.spec.whatwg.org/#dom-messageport-postmessage-options-dev
describes transferable objects as being “no longer usable” on the side
they’re sent from. So we have “becomes unusable” in the docs, and there’s
no reason to introduce any additional term that’s not in the spec.

This is a follow-up to https://github.com/mdn/content/pull/3690, which
made a change to the Channel Messaging API article, for the same reason.

This change also fixes a couple of link/macro flaws.