Minor edits on isNaN()
I encountered this page today, and was confused by the line

> The isNaN() function determines whether a value is NaN or not. Note, coercion inside the isNaN function has interesting rules; you may alternatively want to use Number.isNaN(), as defined in ECMAScript 2015.

Due to the typography, the full stop of `not. Note` became almost impossible to spot. Shall we had a `<div class="notecard note">` here? How would that impact the abstract for SEO? I decided to adjust the wording a bit, to not make it read like `not Note`.

While I was at it, I updated the code samples to drop `var` in favour of `const`.