Normalized Set format and clarified set vs array (#3342)
* Normalized Set format and clarified set vs array

I was reading the docs on this today and was very confused by how and why the docs were doing things. After spending some time testing the examples, I realized that they simply were not consistent in how they were written and were type casting in someplace for no reason. These changes settle on representing the `Set` according to the standard unless explicitly showing how to convert from a set to an array.

This should help devs not already familiar with Set to better understand it and its uniqueness at first glance and reduce confusion like I had by utilizing the standards for Set when applicable

* Update files/en-us/web/javascript/reference/global_objects/set/index.html

Co-authored-by: Ruth John <Rumyra@users.noreply.github.com>

* Update files/en-us/web/javascript/reference/global_objects/set/index.html

Co-authored-by: Ruth John <Rumyra@users.noreply.github.com>

* Update files/en-us/web/javascript/reference/global_objects/set/index.html

Co-authored-by: Ruth John <Rumyra@users.noreply.github.com>

* Update files/en-us/web/javascript/reference/global_objects/set/index.html

added 'Set' to the returns comment for the Set object to increase accuracy.

Co-authored-by: Ruth John <Rumyra@users.noreply.github.com>

Co-authored-by: Ruth John <Rumyra@users.noreply.github.com>