Fixing Parameter in WebXR Code Example
There is a code snippet in the "Rendering and the WebXR frame animation callback" of the WebXR Device API docs that attempts to request a reference space using a session mode value. I have updated it to use `local` as that is the most general but if that is not correct, I am happy to update that to the intended type based on any feedback.