Added hyperlinks to each header on page
There was already a table of contents-like system in place at the top of the page, and I frequent this page a lot. It's annoying to have to scroll all the way down to the 400 or 500 sections, so I am proposing adding hyperlinks to each section.

Here's a preview of what it will look like:

![Changes made locally](https://i.imgur.com/nwMUb5e.png)

I thought it looked a bit off if it included the `(<code>###</code>-<code>###</code>)` in the hyperlink, but feel free to include that if it's preferred.