Note String.p.trim* methods don’t throw if no whitespace trimmed
To the Return Values section of each of the articles for the
String.p.trim* methods, this change adds a statement explaining that if
the string contains no whitespace to trim, no exception is thrown
(instead the methods essentially just return a copy of the string).

Fixes https://github.com/mdn/content/issues/3961#issuecomment-816664813

This change also drops the Description sections from each article —
because the contents of those sections just repeats what’s already
stated in the Return Values section.