Fix caret*FromPoint examples and explanation
This change fixes the examples in the document.caretPositionFromPoint()
and document.caretRangeFromPoint() articles to first try
caretRangeFromPoint() (which is non-standard, but more widely supported)
but then, if that’s not supported, fall back to caretPositionFromPoint().

The change also adds a prose explanation of what the code is doing, as
well we reordering the parts of the example for a better user experience.

Fixes https://github.com/mdn/content/issues/4044