Add an example to the SVG “transform-origin” doc
This change adds an example to the SVG “transform-origin” article. The
example is based on code snippets from a Stack Overflow question (with
attribution give, and with a link to the CC BY-SA license).

The purpose of the example is in part to demonstrate how the same
transformation that can be performed with transform-origin can also be
performed with transform only (which is useful for targeting browsers
that don’t have support for transform-origin — or that don’t support it
as expected).

Related: https://github.com/mdn/browser-compat-data/issues/9834