Clean up documentation about Alt-Svc HTTP header
Some content had been added that described the Alt-Svc header as having
the purpose of “separating the identity and location of a resource”.
But that’s a misleading description — because what the Alt-Svc header is
really about is, allowing a server to indicate that a particular
resource should be loaded from a different server — while still
appearing to the user as if were loaded from the same server.

See https://www.mnot.net/blog/2016/03/09/alt-svc for a great overview.

So this change:

* Replaces the summary in the Alt-Svc article itself, while adding a See
  Also link to that https://www.mnot.net/blog/2016/03/09/alt-svc article.

* Rewrites the description of Alt-Svc in the “Basics of HTTP” article.

* Removes mention of Alt-Svc from the “Identifying resources on the Web”
  article. That article is entirely about URLs and their constituent
  parts; e.g., their path parts. Alt-Svc is not about URLs — it is about
  origins (host/protocol/port) and has nothing do with, e.g., path
  parts. So Alt-Svc doesn’t belong there; it’s a non sequitur there.

* Removes mention of Alt-Svc from the “Resources and URIs” article, for
  reasons similar to the reasons for removing it from the “Identifying
  resources on the Web” article.

Fixes https://github.com/mdn/content/issues/4125