Drop unhelpful note about CSRF from CORS doc
If we’re going to mention CSRF in the CORS article, we would need to say
quite a lot more than what this note says. And since this note is
wrongly saying that you don’t need to worry at all about CSRF in a
specific subset of cases (simple requests), it may give the very wrong
impression that CORS overall doesn’t create any additional CSRF risk.

So it’s better to mention nothing at all about CSRF in this CORS article
than it is to have this one mention in this one odd note.

Fixes https://github.com/mdn/content/issues/4111