Clarify in JS “static” doc: class fields are a proposal
This change adds a statement to the beginning of the article on the
“static” keyword to alert readers to the fact that the examples
throughout the article use public class fields, which are not yet part
of the ES spec but instead are in a proposal document.

Fixes https://github.com/mdn/content/issues/4097

The change also updates the compat table on the page to show support
data for the entire JS Class hierarchy, including public class fields
and static class fields — to help readers understand what the level of
browser support is for the syntax used in the article.

Additionally, the change updates the “Public class fields” and “Private
class fields” articles with See Also links to the following:

* https://v8.dev/features/class-fields
* https://github.com/tc39/proposal-class-fields#class-field-declarations-for-javascript