Switch master to main
Switching over the default branch in the examples to "main" to match current standards. I also wrote a note to explain the switch from "master," which I tried to not make too outspoken while suggesting that main should be the default of choice. By all means feel free to make the note more blunt, and hope this helps :)