Fix information for fullscreenchange event
Fixing wrong statement stating that the that "fullscreenchange" event is sent to the Document object when entire document is entered/exited fullscreen mode. But in real, this event never gets sent directly to document, instead when any Element in the document raises fullscreenchange event, it gets bubbled up to the Document object.