Upgrade h2->h1, h3->h2, and so forth.
Currently there are very few entries in the table of contents sidebar;
just "General Topics", "Shaders, Programs, and GLSL", etc., but there
are many individual pieces of advice. h2 entries show up in the TOC,
so promote everything one heading level to make it easier to see all
of the individual pieces of advice at a glance.