Deprecate PaymentAddress, PayerErrors, and related members
https://github.com/w3c/payment-request/commit/486c07a
(https://github.com/w3c/payment-request/pull/955) removed the
PaymentAddress and PayerErrors objects from the Payment Request
specification — as well as removing related members from the
PaymentRequest, PaymentResponse, and PaymentDetailsUpdate objects.

So this change adds Deprecated and Non-standard banners to the
PaymentAddress and PayerErrors articles, as well as to the articles for
all their member, and to the articles for the related members of the
PaymentRequest, PaymentResponse, and PaymentDetailsUpdate objects.

The change also removes all the Specifications tables for the deprecated
features (they’re no longer in the spec).

Related BCD change: https://github.com/mdn/browser-compat-data/pull/10413