Clean up SVGStylable and SVGLangSpace references
These were mixin-like interfaces in SVG 1.1:
https://www.w3.org/TR/SVG11/types.html#InterfaceSVGStylable
https://www.w3.org/TR/SVG11/types.html#InterfaceSVGLangSpace

The className attribute is already in the SVGElement page, and the style
attributes is in https://developer.mozilla.org/en-US/docs/Web/API/ElementCSSInlineStyle
and will be folded into SVGElement later as part of mixin flattening.

The xmllang and xmlspace attributes are long gone, so there's nothing to
fix there.

All that remains of these is now in the specifications tables.