use getBoundingClientRect to determine progress
getBoundingClientRect should be used to find the left position of the progress bar on the screen instead of offsetLeft. Using offsetLeft was causing the progress position to not be calculated correctly when absolute positioning the progress bar somewhere on the screen. This change lets you position the progress bar anywhere on the screen and still have the calculation work as intended.