DeviceMotionEvent constructor: Relax requirements for the "type" parameter.
There was some confusion in https://crbug.com/1213041 over the previous
text, which said that the `type` parameter must always be "devicemotion".
While it does not make much sense to use a different value, it is possible
to do so, just like with any other Event-derived interface.

Write a more generic description for the parameter based on what we have in
the description of the `MouseEvent` constructor.