text revisions to JS .toLocaleTimeString() doc (#3792)
* Update index.html

syntax corrections in the Syntax section
locales -> locale (except for the ID)
removed Checking for support for locales and options arguments section that was predicated on lack of browser support
grammar and punctuation corrections
other text revisions for clarity and readability

* Update index.html

reverse locales -> locale edits
Note that both the term locale and locales were present in the document before any of my edits. I have reversed only the edits I made for changing locales to locale. Both forms remain present in the text.

Co-authored-by: wbamberg <will@bootbonnet.ca>