html/element/input: Escape dot in regular expression for email
While the `Pattern_validation` example suggested that it checks for
`beststartupever.com` the pattern accepted any character instead of
the dot before `com`. So it matched for example
`foo@beststartupeverZcom`.

While it's harmless in our setting, we want to alert users that they have to
escape dots in regular expressions if they intend to only match dots, we set a
good example and escape the dot before `com`.