Remove events from SVGElement that aren't a part of the spec
This PR removes four events from the SVGElement API: abort, resize, scroll, and unload. None of these events seem to be in the spec specific for SVGElement, and the only data we can find is for the global Window or for GlobalEventHandlers.

This PR is to compliment the removal of these events from BCD in https://github.com/mdn/browser-compat-data/pull/11071.