Give additional examples of how to create a MediaStream.
The current introductory prose makes it sound like there are only two 
ways to create a MediaStream, when there are really others.  This change 
adds two more ways to the list, and also makes the list less firm that 
it's complete.  (I'm not sure whether the ways listed following this 
edit are all the ways that exist today, but even if they are, that seems 
like it may not continue to be true.)