Replace “host” with “hostname” in Origin glossary entry
The term “host” is ambiguous, because as pointed out in
https://github.com/mdn/content/issues/6148, “host” can mean “fully
qualified domain name plus port”. But “hostname” in unambiguous, because
it’s clearly understood to just mean “fully qualified domain name” —
without the port.

So to avoid that any ambiguity in the Origin glossary entry, this change
replaces “host” with “hostname” throughout.

Fixes https://github.com/mdn/content/issues/6148