subtlecrypto: Random key/IV for second AES-CTR encryption example
The second example for AES-CTR encryption initializes a key and IV
as Uint8Array without writing random values to it, leaving it
zero-valued.
The null-value key and IV are directly used for encryption, causing the
encryption operation to be insecure/ineffective.

I suggest to fill key and IV with cryptographically secure random
values. This avoids that developers may copy the snippet from MDN,
leaving IV or key as a null-value.
Especially given that there is no explanation for the second example,
the risk is very high that developers are unaware of the problems of a
non-random/repeatedly used IV or key and directly apply such a snippet.