Clarify 'unsafe-hashes'
This commit adds a sentence to explain an extra action needed when using
the 'unsafe-hashes' attribute.

The reason is this: The name 'unsafe-hashes' implies it has _something_
to do with hashes; but it does not give any clue about having to do with
"inline event handlers". The documentation does though, but the
documentation does not mention what should be hashed. Without this
change, it has to be understood implicitly.