Add .github/workflows/ping-other-repos.yml
For every push to the repo, the .github/workflows/ping-other-repos.yml
file added in this change creates a GitHub repository_dispatch event to
send a ping to the https://github.com/w3c/mdn-spec-links/ repo (where it
triggers the build of the contents of that repo).

Along with https://github.com/w3c/mdn-spec-links/, other repos can be
added to this later — any repos that want to have some action triggered
each time a push to the BCD repo occurs.

https://github.com/marketplace/actions/repository-dispatch has docs on the
GitHub Action this uses, and docs for the underlying API it uses are at
https://docs.github.com/en/rest/reference/repos#create-a-repository-dispatch-event