"?." instead of "Optional chaining" in ref title
Some people (like myself) are not familiar from the start with the general operator names, especially for the (relatively) new ones. Personally I have looked for ?. in the list of all operators on this page and could not find it quite for a while not knowing that its "right" name is "optional chaining". Therefore I hope that putting the operator itself as the title into the operator list will be more readable, especially for javascript learners.