DeviceMotionEvent: Improve explanation of acceleration attributes. (#6675)
`accelerationIncludingGravity` is a bad name that we unfortunately have to
live with (see w3c/deviceorientation#96). It essentially represents raw
accelerometer data, whose values can be surprising: when a device is at rest
with the screen pointing up, it reports +9.8 in the Z axis rather than 0;
when the device is in free fall, it reports 0 in the Z axis.

Try to explain this in a succint way while also pointing to more detailed
information in related Wikipedia articles. The changes also change a
sentence that said the attribute represented a sum of acceleration on the
device and acceleration caused by gravity. This is confusing because the
DeviceOrientation Event spec defines `accelerationIncludingGravity` as
acceleration on the device plus an acceleration equal and _opposite_ to
gravity.

While here, be clear that the `accelerometer` attribute does not include the
effects of gravity, and reference the corresponding Generic Sensor API
interfaces that provide the same data in an arguably better way and with
better names: `Accelerometer` and `LinearAccelerationSensor`.