Un-link the word “instantiating” in “Working with objects” doc
This change removes a broken link to an anchor for an article section
which no longer exists. It’s anyway an unnecessary link, since we use
the word “instantiat(e|ing)” in plenty of other places in the docs,
without it being hypertext.

Fixes https://github.com/mdn/content/issues/6696