Updating examples in "Inheritance and the prototype chain" (#6803)
* Updating examples in "Inheritance and the prototype chain"

There were few instances where a new object was assigned to the prototype property of a function. This is not a good approach. Instead, in order to add a new property, it should be added to the existing prototype of the function.
For example, 
foo.prototype = {
  foo_prop: "foo val"
};
should be
foo.prototype.foo_prop = "foo val";

Also, in some places, the constructor function was not being called while being initialized with the new keyword.

* Removing redundant code