Drop duplicated note from “JavaScript guidelines” doc
This note is a non sequitur in this context, since the context here is
named functions, and the note cites examples of variable names — and
also because unlike the case of i, j, etc. for loop iterators, there is
no analogous case for named functions; that is, there is no common
practice of ever using non-“human-readable” names for named functions.

https://github.com/mdn/content/issues/6847