added clarity
The concept of a "border box" is a novel one. I have at least some understanding of what "border box" is as a value of "box-sizing" because I already understand the concepts of properties and values. But in the last sentence, "By doing this you are telling the browser to take the border box as the area defined by any size you set." The concept of border box is reinvoked, but this time not as a value. So the meaning seems to have changed. If so, its meaning should be clarified. For example, does "border box" in the sentence equate to "alternative CSS box"?