added clarity
Also, I've never seen a present participle or a gerund used to signify the name of a categorial concept before. It sounds odd to say "margin collapsing" as if it were such a concept. I may very well be wrong here, since my skills are with English and not with web development.