Update on let and const variable on hoisting.
I update on let and const variable on hoisting behavior. As your page said that let and const are not hoisted but they hoisted in a temporal dead zone so whenever we try to access the let and const variable in a temporal dead zone it gives us a Reference Error and after assigning the values to variables the temporal dead zone will vanish.