Adding an Active Learning Section
Active Learning
The next step after Project Ideation is to map out a workable learning strategy. 
Learning is never an easy activity, whether it’s something new or old. Irrespective of what you are trying to learn, commitment is a key prerequisite to learning actively. If you are not committed, you can be easily distracted.
The same thing applies if you are trying to design your website. Because learning is a process, you need to understand that the curve is steep; that there are highs and lows and sometimes, you will feel like giving up. This is where active learning helps push your forward towards the achievement of your end goal. It also helps keep you on track and you know that with each additional time you spend trying to accomplish your goal, the closer you are.
How then can you learn actively? You must be involved. Basically, that is it. Whether you are reading a written content, like MDN documentation or watching a video tutorial, merely trying to take whatever new information you are exposed to will not help you, you need to participate. You need to try things on your own (put in a list). 
If you are learning HTML tables, for instance, after seeing the instructor’s solution, you need to try it on your own and then try to improve on it by adding your own ideas. Let’s say your lecture material has an HTML table with 2 rows and 2 columns, you can implement this first and then go further to implement a table with 3 rows and 2 columns, 5 rows and 7 columns and so on. By so doing, it becomes easier to retain whatever new concepts you have learnt.