Drop the Implementations list from “JavaScript language resources”
The Implementations list includes three links to MDN articles for
SpiderMonkey, Rhino, and Tamarin that are all now 404 because those
articles are long-gone. Since there’s no information under this heading,
but instead just a list with those broken links, it’s in no way
essential to the article, so let’s just drop it.

Relates to https://github.com/mdn/content/issues/7444