Clean up some code examples
This pr fixes the code in the examples for attaching a reader. Previously the code was not complete useable code, which may confuse inexperienced users who try and copy-paste this code and run it. 
I'm not sure if this is the source code or auto-generated from markdown or something, and I'd be glad to make a pr on the source if someone can tell me where it is.