Replace Text object description with clear example
In https://developer.mozilla.org/en-US/docs/Web/API/Text the current
summary/description of the Text interface was copied verbatim from
https://www.w3.org/TR/2004/REC-DOM-Level-3-Core-20040407/core.html#ID-1312295772.
And that description is really not actually clear or useful to developers.

So, this change replaces that description with an example that makes
clear what a text node is.

Fixes https://github.com/mdn/content/issues/5487