Refine the Array.prototype.reduce() doc for clarity
We have multiple open issues — #6156 and #6497 — reporting that the
current description we have in MDN for how Array.prototype.reduce()
works is confusing, and asking that we refine it to make it more clear.

So this change replaces the current description by citing the
https://tc39.es/ecma262/#sec-array.prototype.reduce description in the
spec — which is much more clear — and then updating related text and
variable names to match the spec language.

One key difference that helps make the spec language much more clear is
that it doesn’t use the word “accumulator” but instead uses
“previousValue”. So for consistency, this change also replaces
“accumulator” throughout with “previousValue".

Fixes https://github.com/mdn/content/issues/6156
Fixes https://github.com/mdn/content/issues/6497

The change also drops the entire Polyfill section — in keeping with our
current policy of not any longer maintaining polyfill code in MDN
articles, but instead linking to polyfills at external sites (and in
this case, the article already has a link to a high-quality polyfill).