Fix mixin example in Object.assign
The previous version was mutating the first superclass's prototype (i.e. `Object.getPrototypeOf(MyClass.prototype)` returns `SuperClass.prototype`, and we don't want to mutate that).

That said, in my opinion this is a horrifically ugly pattern and should never be used, so I'd also be happy to just delete the whole section.