Increase live-output height of “Getting started with HTML” example
This change increases the height of the “This is my page” embedded live
sample in the “Getting started with HTML” article.

Otherwise, without this change, when the reader pushes the “Show solution”
button, the “Editable code” area appears to be replaced by a large image.
However, in fact, the “Editable code” area isn’t replaced by the image;
instead the image gets placed vertically before the “Editable code” area.
And because the height of the iframe for the embedded live sample is too
small, the “Editable code” area seems to disappear. In fact, though, the
“Editable code” area is still there, but the reader needs to scroll the
iframe to view it — which is not immediately obvious to most readers.

So this change increases the height of the iframe for the embedded live
sample such that the “Editable code” area doesn’t get cut off — and
seemingly disappear — when the reader presses the “Show solution” button.

Fixes https://github.com/mdn/content/issues/7883