Align the “Object” glossary entry with formal definition
This change rewrites the “Object” glossary entry to largely just follow
the more-accurate JavaScript-object-as-data-structure explanation at
https://developer.mozilla.org/en-US/docs/Web/JavaScript/Data_structures#objects
and to help preempt reader confusion about whether JavaScript objects
are classes.

Otherwise, the glossary entry as currently written is misleading —
because in so many words it’s basically describing classes, not objects;
and it’s also not really very helpful to readers — because it’s not
describing anything notable about *JavaScript* objects, as opposed to
objects in other programming languages.

And given that MDN is a technical resource for developers working on a
client platform whose programming language is JavaScript, it makes sense
for the glossary to not just have a generic description of “objects” but
to instead have a slightly more technical (but still succinct)
description that’s specific to JavaScript objects and that’s framed in
terms of their characteristics as data structures within the language.

Fixes https://github.com/mdn/content/issues/7895