Drop “Success stories” section of “Intro to progressive web apps”
The “Success stories” section of the “Introduction to progressive web
apps” feels very dated, like reading a list of what the flagship PWA
sites were 5 years ago — which is pretty much exactly what it is, so
it’s not surprising it reads that way.

Among the problems with creating such a list of sites to begin with is
that it invites others to do self-promotion by adding their own sites to
the list — as in https://github.com/mdn/content/pull/7912. And even if
it weren’t a vector for self-promotion and spam, such site lists
generally end up rotting over time, because nobody has any interest in
actively maintaining/updating/pruning them after they’re created.

For those reasons, this change removes the entire “Success stories”
section of the “Introduction to progressive web apps” article.

(And hopefully our current review mechanism will cause any further site
lists like that one to not be added again to the docs to begin with...)