Streamline JS “Data and Structure types” to align with spec
This change strips down the existing “Data and Structure types” section
of the “JavaScript data types and data structures” article to a simple
hyperlinked listing that aligns with the corresponding ES spec section
at https://tc39.es/ecma262/#sec-ecmascript-language-types; notably:

* Changes the heading to “JavaScript types”. (The ES spec section is
  titled “ECMAScript Language Types” but the “Language” is there simply
  to distinguish it from what the ES calls “Specification Types” — which
  are spec-internal types that aren’t actually exposed to developers.)

* Reduces the high-level taxonomy to just *objects* and *primitive values*.

Otherwise, without this change, the current contents of the section read
like a non-sequitur — in that the current section:

* Introduces a taxonomy and terminology not used in the spec (“Structural
  types”, “Structural root primitive”, “non-data structure”).

* Digresses into an (over)focus on what the corresponding typeof
  information is for each type — which is simply an unnecessary
  repetition of what we already have in typeof article itself at
  https://developer.mozilla.org/docs/Web/JavaScript/Reference/Operators/typeof

The existing section as a whole kind of reads like one of the idiosyncratic
additions that got added back in the wiki days by well-intentioned contributors
but that never got the kind of scrutiny and technical review they should’ve had.

So this change strips this all back down to just what’s useful and necessary
and consistent with the spec. If it turns out that more needs to be said
in this section, we can build back up on it more later, starting from this.

Fixes https://github.com/mdn/content/issues/7939