Fix reference in bold in <pre> in “Django introduction” doc
In Markdown we don’t have the ability to make parts of code blocks bold
(or italic either). So to help make the “Django introduction” article
Markdown-ready, the existing “The line in bold [in the code block below]
shows” statement needs to be changed to make sense without requiring
boldfacing of the relevant line in the code block. The fix that’s
easiest and clearest is to just replace the “The line in bold” reference
by copying the relevant line verbatim into the paragraph.

Fixes https://github.com/mdn/content/issues/7958