Mention in main CORS doc how wildcards work w/ ACAH & ACAM headers
This change updates the “Credentialed requests and wildcards” section of
the https://developer.mozilla.org/en-US/docs/Web/HTTP/CORS article to
mention that the "*" wildcard must not be used in the values of the
Access-Control-Allow-Headers and Access-Control-Allow-Methods headers in
responses to credentialed requests.

The Access-Control-Allow-Headers & Access-Control-Allow-Methods articles
themselves already very clearly mention the restriction on not using the
"*" wildcard in responses to credentialed requests — but it’s also
helpful to developers to repeat that info in the main CORS article.

https://stackoverflow.com/q/68793536/ is a recent Stack Overflow
question from a developer confused by not finding that restriction
mentioned in the main CORS article. So that’s probably a good indication
that there are number of other developers out there who end up being
similarly confused after reading only the main CORS article but not the
Access-Control-Allow-Headers & Access-Control-Allow-Methods articles.