Radically streamline the PWA landing page
https://developer.mozilla.org/en-US/docs/Web/Progressive_web_apps is a
landing/portal page that should just serve as a gateway to attract users
into the reading the subpages where all the useful content actually is.

But that existing page has so much content on it — and so much clutter —
that it’s not serving its purpose well at all.

So this change pares that down to just a well-structured, slim list of links
for quickly getting readers over to the where the meat of the content is.

Among the changes:

- Drop huge PWA logo taking up 25% of screen real estate above the fold
- Copy-edit/rewrite intro paragraphs down to just two sentences
- Drop link redundancy/wordiness caused by {{LandingPageListSubpages}} macro
- Assemble links to subpages into a better order/structure for readers
- Discard cruft from list of links to external sites
- Distill external-sites list down to a curated set of just the best stuff

Fixes https://github.com/mdn/content/issues/8023