Drop outdated/404 links from “JavaScript shells” doc
This change drops a number of links in the “JavaScript shells” doc that
are now 404 or are outdated/irrelevant to the degree that we don’t want
to be pointing developers to them any longer.

The change also updates a few http URLs to https.

Fixes https://github.com/mdn/content/issues/8085