Drop/replace outdated links in “Firefox” page
All the links dropped or replaced in this change are for source that
moved to the archived-content repo and so that have now been deleted
from the production site.

The only replacement is for “mozilla-central” — changed to a direct link
to the Mercurial repo (rather than the intro page that had been in MDN).

Fixes https://github.com/mdn/content/issues/8293